+++ cookbooks/docker/libraries/helpers_service.rb
@@ -126,10 +126,14 @@ module DockerCookbook
         opts << "--bridge=#{bridge}" if bridge
         opts << "--bip=#{bip}" if bip
         opts << '--debug' if debug
+        opts << "--cluster-advertise=#{cluster_advertise}" if cluster_advertise
+        opts << "--cluster-store=#{cluster_store}" if cluster_store
+        cluster_store_opts.each { |store_opt| opts << "--cluster-store-opt=#{store_opt}" } if cluster_store_opts
         default_ulimit.each { |u| opts << "--default-ulimit=#{u}" } if default_ulimit
         dns.each { |dns| opts << "--dns=#{dns}" } if dns
         dns_search.each { |dns| opts << "--dns-search=#{dns}" } if dns_search
         opts << "--exec-driver=#{exec_driver}" if exec_driver
+        exec_opts.each { |exec_opt| opts << "--exec-opt=#{exec_opt}" } if exec_opts
         opts << "--fixed-cidr=#{fixed_cidr}" if fixed_cidr
         opts << "--fixed-cidr-v6=#{fixed_cidr_v6}" if fixed_cidr_v6
         opts << "--group=#{group}" if group