+++ cookbooks/docker_wrapper/resources/container_service.rb
@@ -1,0 +1,5 @@
+actions :start, :enable, :disable, :stop
+default_action :start
+
+attribute :service, :kind_of => [String], :name_attribute => true
+attribute :container_name, :kind_of => [String]