+++ cookbooks/packer_docker/kvm_docker-server_dependencies/metadata.rb
@@ -1,0 +1,8 @@
+name             'kvm_docker-server_dependencies'
+maintainer       'YOUR_NAME'
+maintainer_email 'YOUR_EMAIL'
+license          'All rights reserved'
+description      'Installs/Configures kvm_docker-server_dependencies'
+long_description 'Installs/Configures kvm_docker-server_dependencies'
+version          '0.1.0'
+depends          'chef-client'